added avoidance if same dish name

+++ src/components/NewRecipeFormContainer.js
@@ -10,20 +10,31 @@ class Form extends React.Component {
         let newIngredients = document.getElementById("newIngredients").value;
         let newInstructions = document.getElementById("newInstructions").value;
         let newImage = document.getElementById("newImage").value;
-        let newRecipe = {
-            name: newDishName,
-            ingredients: newIngredients.split(","),
-            instructions: newInstructions.split("."),
-            image: newImage
-        }
         let allRecipes = JSON.parse(window.localStorage.getItem("recipeStorage"));
-        let newRecipeList = [];
+        let dishNameExists = false;
         for (let i = 0; i < allRecipes.length; i += 1) {
-            newRecipeList.push(allRecipes[i]);
+            if (allRecipes[i].name == newDishName) {
+                dishNameExists = true;
+            }
         }
-        newRecipeList.push(newRecipe);
-        window.localStorage.setItem("recipeStorage", JSON.stringify(newRecipeList));
+        if (dishNameExists) {
+            alert("Dish already exists. Please choose another name.");
+        } else {
+            let newRecipe = {
+                name: newDishName,
+                ingredients: newIngredients.split(","),
+                instructions: newInstructions.split("."),
+                image: newImage
+            }
+            let newRecipeList = [];
+            for (let i = 0; i < allRecipes.length; i += 1) {
+                newRecipeList.push(allRecipes[i]);
+            }
+            newRecipeList.push(newRecipe);
+            window.localStorage.setItem("recipeStorage", JSON.stringify(newRecipeList));
         
+
+        }
     }
 
     render() {
